more changes to assets page
Related Issue: SPBK-4186
