+++ ui/src/components/routes/cases/detail/assets/Asset.tsx
@@ -3,28 +3,7 @@ import type { Case } from 'models/entities/generated/Case';
 import type { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
-
-export type AssetType = 'hash' | 'hosts' | 'ip' | 'user' | 'ids' | 'id' | 'uri' | 'signature';
-
-export type AssetRole = 'threat' | 'target' | 'indicator';
-
-export interface AssetSource {
-  id: string;
-  type: 'hit' | 'observable' | 'case';
-  path?: string;
-  escalation?: string;
-}
-
-export interface AssetEntry {
-  type: AssetType;
-  value: string;
-  /** IDs of the hits/observables this asset was seen in */
-  seenIn: string[];
-  /** Resolved source metadata for each seenIn item */
-  sources?: AssetSource[];
-  /** Classified role of this asset */
-  role?: AssetRole;
-}
+import type { AssetEntry } from '../types';
 
 const Asset: FC<{ asset: AssetEntry; case: Case }> = ({ asset, case: _case }) => {
   const { t } = useTranslation();
